# docs/swagger.json
{
    "schemes": [
        "http",
        "https"
    ],
    "swagger": "2.0",
    "info": {
        "description": "High-Performance Anime Streaming API built with Go\nA blazing-fast, production-ready RESTful API for anime streaming platforms\n\n**Features:**\n- Complete anime data (episodes, metadata, streaming sources)\n- Advanced search and filtering\n- Schedule management and notifications\n- Character and voice actor information\n- User watchlist support\n- Optimized caching with Redis\n\n**Disclaimer:**\nThis API does not store any files. It only links to media hosted on 3rd party services.\nThis API is explicitly made for educational purposes only.",
        "title": "Jutsu API",
        "contact": {
            "name": "Jutsu API Support",
            "url": "https://github.com/mnuddindev/jutsu-api",
            "email": "support@jutsu-api.com"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0.0"
    },
    "host": "localhost:8080",
    "basePath": "/api",
    "paths": {
...
        "/az-list": {
            "get": {
                "description": "Returns paginated anime rows for a given category or genre route (e.g. /api/genre/action, /api/top-airing)",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Categories"
                ],
                "summary": "Get category or genre listing",
                "parameters": [
                    {
                        "type": "integer",
                        "default": 1,
                        "example": 1,
                        "description": "Page number (must be \u003e= 1)",
                        "name": "page",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Category results",
                        "schema": {
                            "$ref": "#/definitions/handler.CategoryResponse"
                        }
                    },
                    "400": {
                        "description": "Invalid page parameter",
                        "schema": {
                            "$ref": "#/definitions/handler.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Requested page exceeds total available pages",
                        "schema": {
                            "$ref": "#/definitions/handler.ErrorResponse"
                        }
                    },
                    "502": {
                        "description": "Failed to load category data",
                        "schema": {
                            "$ref": "#/definitions/handler.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/az-list/{letter}": {
            "get": {
                "description": "Returns paginated anime rows for a given category or genre route (e.g. /api/genre/action, /api/top-airing)",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Categories"
                ],
                "summary": "Get category or genre listing",
                "parameters": [
                    {
                        "type": "integer",
                        "default": 1,
                        "example": 1,
                        "description": "Page number (must be \u003e= 1)",
                        "name": "page",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Category results",
                        "schema": {
                            "$ref": "#/definitions/handler.CategoryResponse"
                        }
                    },
                    "400": {
                        "description": "Invalid page parameter",
                        "schema": {
                            "$ref": "#/definitions/handler.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Requested page exceeds total available pages",
                        "schema": {
                            "$ref": "#/definitions/handler.ErrorResponse"
                        }
                    },
                    "502": {
                        "description": "Failed to load category data",
                        "schema": {
                            "$ref": "#/definitions/handler.ErrorResponse"
                        }
                    }
                }
            }
        },
...
        "/completed": {
            "get": {
                "description": "Returns paginated anime rows for a given category or genre route (e.g. /api/genre/action, /api/top-airing)",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Categories"
                ],
                "summary": "Get category or genre listing",
                "parameters": [
                    {
                        "type": "integer",
                        "default": 1,
                        "example": 1,
                        "description": "Page number (must be \u003e= 1)",
                        "name": "page",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Category results",
                        "schema": {
                            "$ref": "#/definitions/handler.CategoryResponse"
                        }
                    },
                    "400": {
                        "description": "Invalid page parameter",
                        "schema": {
                            "$ref": "#/definitions/handler.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Requested page exceeds total available pages",
                        "schema": {
                            "$ref": "#/definitions/handler.ErrorResponse"
                        }
                    },
                    "502": {
                        "description": "Failed to load category data",
                        "schema": {
                            "$ref": "#/definitions/handler.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/dubbed-anime": {
            "get": {
                "description": "Returns paginated anime rows for a given category or genre route (e.g. /api/genre/action, /api/top-airing)",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Categories"
                ],
                "summary": "Get category or genre listing",
                "parameters": [
                    {
                        "type": "integer",
                        "default": 1,
                        "example": 1,
                        "description": "Page number (must be \u003e= 1)",
                        "name": "page",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Category results",
                        "schema": {
                            "$ref": "#/definitions/handler.CategoryResponse"
                        }
                    },
                    "400": {
                        "description": "Invalid page parameter",
                        "schema": {
                            "$ref": "#/definitions/handler.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Requested page exceeds total available pages",
                        "schema": {
                            "$ref": "#/definitions/handler.ErrorResponse"
                        }
                    },
                    "502": {
                        "description": "Failed to load category data",
                        "schema": {
                            "$ref": "#/definitions/handler.ErrorResponse"
                        }
                    }
                }
            }
        },
...
        "/genre/{slug}": {
            "get": {
                "description": "Returns paginated anime rows for a given category or genre route (e.g. /api/genre/action, /api/top-airing)",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Categories"
                ],
                "summary": "Get category or genre listing",
                "parameters": [
                    {
                        "type": "integer",
                        "default": 1,
                        "example": 1,
                        "description": "Page number (must be \u003e= 1)",
                        "name": "page",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Category results",
                        "schema": {
                            "$ref": "#/definitions/handler.CategoryResponse"
                        }
                    },
                    "400": {
                        "description": "Invalid page parameter",
                        "schema": {
                            "$ref": "#/definitions/handler.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Requested page exceeds total available pages",
                        "schema": {
                            "$ref": "#/definitions/handler.ErrorResponse"
                        }
                    },
                    "502": {
                        "description": "Failed to load category data",
                        "schema": {
                            "$ref": "#/definitions/handler.ErrorResponse"
                        }
                    }
                }
            }
        },
...
        "/most-favorite": {
            "get": {
                "description": "Returns paginated anime rows for a given category or genre route (e.g. /api/genre/action, /api/top-airing)",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Categories"
                ],
                "summary": "Get category or genre listing",
                "parameters": [
                    {
                        "type": "integer",
                        "default": 1,
                        "example": 1,
                        "description": "Page number (must be \u003e= 1)",
                        "name": "page",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Category results",
                        "schema": {
                            "$ref": "#/definitions/handler.CategoryResponse"
                        }
                    },
                    "400": {
                        "description": "Invalid page parameter",
                        "schema": {
                            "$ref": "#/definitions/handler.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Requested page exceeds total available pages",
                        "schema": {
                            "$ref": "#/definitions/handler.ErrorResponse"
                        }
                    },
                    "502": {
                        "description": "Failed to load category data",
                        "schema": {
                            "$ref": "#/definitions/handler.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/most-popular": {
            "get": {
                "description": "Returns paginated anime rows for a given category or genre route (e.g. /api/genre/action, /api/top-airing)",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Categories"
                ],
                "summary": "Get category or genre listing",
                "parameters": [
                    {
                        "type": "integer",
                        "default": 1,
                        "example": 1,
                        "description": "Page number (must be \u003e= 1)",
                        "name": "page",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Category results",
                        "schema": {
                            "$ref": "#/definitions/handler.CategoryResponse"
                        }
                    },
                    "400": {
                        "description": "Invalid page parameter",
                        "schema": {
                            "$ref": "#/definitions/handler.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Requested page exceeds total available pages",
                        "schema": {
                            "$ref": "#/definitions/handler.ErrorResponse"
                        }
                    },
                    "502": {
                        "description": "Failed to load category data",
                        "schema": {
                            "$ref": "#/definitions/handler.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/movie": {
            "get": {
                "description": "Returns paginated anime rows for a given category or genre route (e.g. /api/genre/action, /api/top-airing)",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Categories"
                ],
                "summary": "Get category or genre listing",
                "parameters": [
                    {
                        "type": "integer",
                        "default": 1,
                        "example": 1,
                        "description": "Page number (must be \u003e= 1)",
                        "name": "page",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Category results",
                        "schema": {
                            "$ref": "#/definitions/handler.CategoryResponse"
                        }
                    },
                    "400": {
                        "description": "Invalid page parameter",
                        "schema": {
                            "$ref": "#/definitions/handler.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Requested page exceeds total available pages",
                        "schema": {
                            "$ref": "#/definitions/handler.ErrorResponse"
                        }
                    },
                    "502": {
                        "description": "Failed to load category data",
                        "schema": {
                            "$ref": "#/definitions/handler.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/ona": {
            "get": {
                "description": "Returns paginated anime rows for a given category or genre route (e.g. /api/genre/action, /api/top-airing)",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Categories"
                ],
                "summary": "Get category or genre listing",
                "parameters": [
                    {
                        "type": "integer",
                        "default": 1,
                        "example": 1,
                        "description": "Page number (must be \u003e= 1)",
                        "name": "page",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Category results",
                        "schema": {
                            "$ref": "#/definitions/handler.CategoryResponse"
                        }
                    },
                    "400": {
                        "description": "Invalid page parameter",
                        "schema": {
                            "$ref": "#/definitions/handler.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Requested page exceeds total available pages",
                        "schema": {
                            "$ref": "#/definitions/handler.ErrorResponse"
                        }
                    },
                    "502": {
                        "description": "Failed to load category data",
                        "schema": {
                            "$ref": "#/definitions/handler.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/ova": {
            "get": {
                "description": "Returns paginated anime rows for a given category or genre route (e.g. /api/genre/action, /api/top-airing)",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Categories"
                ],
                "summary": "Get category or genre listing",
                "parameters": [
                    {
                        "type": "integer",
                        "default": 1,
                        "example": 1,
                        "description": "Page number (must be \u003e= 1)",
                        "name": "page",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Category results",
                        "schema": {
                            "$ref": "#/definitions/handler.CategoryResponse"
                        }
                    },
                    "400": {
                        "description": "Invalid page parameter",
                        "schema": {
                            "$ref": "#/definitions/handler.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Requested page exceeds total available pages",
                        "schema": {
                            "$ref": "#/definitions/handler.ErrorResponse"
                        }
                    },
                    "502": {
                        "description": "Failed to load category data",
                        "schema": {
                            "$ref": "#/definitions/handler.ErrorResponse"
                        }
                    }
                }
            }
        },
...
        "/recently-added": {
            "get": {
                "description": "Returns paginated anime rows for a given category or genre route (e.g. /api/genre/action, /api/top-airing)",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Categories"
                ],
                "summary": "Get category or genre listing",
                "parameters": [
                    {
                        "type": "integer",
                        "default": 1,
                        "example": 1,
                        "description": "Page number (must be \u003e= 1)",
                        "name": "page",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Category results",
                        "schema": {
                            "$ref": "#/definitions/handler.CategoryResponse"
                        }
                    },
                    "400": {
                        "description": "Invalid page parameter",
                        "schema": {
                            "$ref": "#/definitions/handler.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Requested page exceeds total available pages",
                        "schema": {
                            "$ref": "#/definitions/handler.ErrorResponse"
                        }
                    },
                    "502": {
                        "description": "Failed to load category data",
                        "schema": {
                            "$ref": "#/definitions/handler.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/recently-updated": {
            "get": {
                "description": "Returns paginated anime rows for a given category or genre route (e.g. /api/genre/action, /api/top-airing)",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Categories"
                ],
                "summary": "Get category or genre listing",
                "parameters": [
                    {
                        "type": "integer",
                        "default": 1,
                        "example": 1,
                        "description": "Page number (must be \u003e= 1)",
                        "name": "page",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Category results",
                        "schema": {
                            "$ref": "#/definitions/handler.CategoryResponse"
                        }
                    },
                    "400": {
                        "description": "Invalid page parameter",
                        "schema": {
                            "$ref": "#/definitions/handler.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Requested page exceeds total available pages",
                        "schema": {
                            "$ref": "#/definitions/handler.ErrorResponse"
                        }
                    },
                    "502": {
                        "description": "Failed to load category data",
                        "schema": {
                            "$ref": "#/definitions/handler.ErrorResponse"
                        }
                    }
                }
            }
        },
...
        "/special": {
            "get": {
                "description": "Returns paginated anime rows for a given category or genre route (e.g. /api/genre/action, /api/top-airing)",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Categories"
                ],
                "summary": "Get category or genre listing",
                "parameters": [
                    {
                        "type": "integer",
                        "default": 1,
                        "example": 1,
                        "description": "Page number (must be \u003e= 1)",
                        "name": "page",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Category results",
                        "schema": {
                            "$ref": "#/definitions/handler.CategoryResponse"
                        }
                    },
                    "400": {
                        "description": "Invalid page parameter",
                        "schema": {
                            "$ref": "#/definitions/handler.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Requested page exceeds total available pages",
                        "schema": {
                            "$ref": "#/definitions/handler.ErrorResponse"
                        }
                    },
                    "502": {
                        "description": "Failed to load category data",
                        "schema": {
                            "$ref": "#/definitions/handler.ErrorResponse"
                        }
                    }
                }
            }
        },
...
        "/subbed-anime": {
            "get": {
                "description": "Returns paginated anime rows for a given category or genre route (e.g. /api/genre/action, /api/top-airing)",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Categories"
                ],
                "summary": "Get category or genre listing",
                "parameters": [
                    {
                        "type": "integer",
                        "default": 1,
                        "example": 1,
                        "description": "Page number (must be \u003e= 1)",
                        "name": "page",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Category results",
                        "schema": {
                            "$ref": "#/definitions/handler.CategoryResponse"
                        }
                    },
                    "400": {
                        "description": "Invalid page parameter",
                        "schema": {
                            "$ref": "#/definitions/handler.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Requested page exceeds total available pages",
                        "schema": {
                            "$ref": "#/definitions/handler.ErrorResponse"
                        }
                    },
                    "502": {
                        "description": "Failed to load category data",
                        "schema": {
                            "$ref": "#/definitions/handler.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/top-airing": {
            "get": {
                "description": "Returns paginated anime rows for a given category or genre route (e.g. /api/genre/action, /api/top-airing)",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Categories"
                ],
                "summary": "Get category or genre listing",
                "parameters": [
                    {
                        "type": "integer",
                        "default": 1,
                        "example": 1,
                        "description": "Page number (must be \u003e= 1)",
                        "name": "page",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Category results",
                        "schema": {
                            "$ref": "#/definitions/handler.CategoryResponse"
                        }
                    },
                    "400": {
                        "description": "Invalid page parameter",
                        "schema": {
                            "$ref": "#/definitions/handler.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Requested page exceeds total available pages",
                        "schema": {
                            "$ref": "#/definitions/handler.ErrorResponse"
                        }
                    },
                    "502": {
                        "description": "Failed to load category data",
                        "schema": {
                            "$ref": "#/definitions/handler.ErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "handler.CategoryResponse": {
            "type": "object",
            "properties": {
                "results": {
                    "type": "object",
                    "properties": {
                        "data": {},
                        "totalPages": {
                            "type": "integer",
                            "example": 10
                        }
                    }
                },
                "success": {
                    "type": "boolean",
                    "example": true
                }
            }
        },
...
        "handler.ErrorResponse": {
            "description": "Standard error response",
            "type": "object",
            "properties": {
                "message": {
                    "type": "string",
                    "example": "Error message here"
                },
                "success": {
                    "type": "boolean",
                    "example": false
                }
            }
        },
...
    },
...
}